+++ src/test/java/org/opensearch/knn/index/query/exactsearch/ByteVectorIdsExactKNNIteratorTests.java
@@ -3,17 +3,18 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package org.opensearch.knn.index.query.iterators;
+package org.opensearch.knn.index.query.exactsearch;
 
+import junit.framework.TestCase;
 import lombok.SneakyThrows;
 import org.apache.lucene.search.DocIdSetIterator;
 import org.apache.lucene.util.BitSetIterator;
 import org.apache.lucene.util.FixedBitSet;
 import org.mockito.stubbing.OngoingStubbing;
-import org.opensearch.knn.KNNTestCase;
 import org.opensearch.knn.index.SpaceType;
-import org.opensearch.knn.index.vectorvalues.KNNFloatVectorValues;
+import org.opensearch.knn.index.vectorvalues.KNNByteVectorValues;
 
+import java.io.IOException;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -24,22 +25,19 @@ import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-public class VectorIdsKNNIteratorTests extends KNNTestCase {
+public class ByteVectorIdsExactKNNIteratorTests extends TestCase {
     @SneakyThrows
-    public void testNextDoc_whenCalledWithFilters_thenIterateAllDocs() {
+    public void testNextDoc_whenCalled_IterateAllDocs() {
         final SpaceType spaceType = SpaceType.L2;
-        final float[] queryVector = { 1.0f, 2.0f, 3.0f };
+        final byte[] byteQueryVector = { 1, 2, 3 };
+        final float[] queryVector = { 1f, 2f, 3f };
         final int[] filterIds = { 1, 2, 3 };
-        final List<float[]> dataVectors = Arrays.asList(
-            new float[] { 11.0f, 12.0f, 13.0f },
-            new float[] { 14.0f, 15.0f, 16.0f },
-            new float[] { 17.0f, 18.0f, 19.0f }
-        );
+        final List<byte[]> dataVectors = Arrays.asList(new byte[] { 11, 12, 13 }, new byte[] { 14, 15, 16 }, new byte[] { 17, 18, 19 });
         final List<Float> expectedScores = dataVectors.stream()
-            .map(vector -> spaceType.getKnnVectorSimilarityFunction().compare(queryVector, vector))
+            .map(vector -> spaceType.getKnnVectorSimilarityFunction().compare(byteQueryVector, vector))
             .collect(Collectors.toList());
 
-        KNNFloatVectorValues values = mock(KNNFloatVectorValues.class);
+        KNNByteVectorValues values = mock(KNNByteVectorValues.class);
         when(values.getVector()).thenReturn(dataVectors.get(0), dataVectors.get(1), dataVectors.get(2));
 
         FixedBitSet filterBitSet = new FixedBitSet(4);
@@ -49,7 +47,7 @@ public class VectorIdsKNNIteratorTests extends KNNTestCase {
         }
 
         // Execute and verify
-        VectorIdsKNNIterator iterator = new VectorIdsKNNIterator(
+        ByteVectorIdsExactKNNIterator iterator = new ByteVectorIdsExactKNNIterator(
             new BitSetIterator(filterBitSet, filterBitSet.length()),
             queryVector,
             values,
@@ -63,21 +61,22 @@ public class VectorIdsKNNIteratorTests extends KNNTestCase {
     }
 
     @SneakyThrows
-    public void testNextDoc_whenCalledWithoutFilters_thenIterateAllDocs() {
+    public void testNextDoc_whenCalled_thenIterateAllDocsWithoutFilter() throws IOException {
         final SpaceType spaceType = SpaceType.L2;
+        final byte[] byteQueryVector = { 1, 2, 3 };
         final float[] queryVector = { 1.0f, 2.0f, 3.0f };
-        final List<float[]> dataVectors = Arrays.asList(
-            new float[] { 11.0f, 12.0f, 13.0f },
-            new float[] { 14.0f, 15.0f, 16.0f },
-            new float[] { 17.0f, 18.0f, 19.0f },
-            new float[] { 20.0f, 21.0f, 22.0f },
-            new float[] { 23.0f, 24.0f, 25.0f }
+        final List<byte[]> dataVectors = Arrays.asList(
+            new byte[] { 11, 12, 13 },
+            new byte[] { 14, 15, 16 },
+            new byte[] { 17, 18, 19 },
+            new byte[] { 20, 21, 22 },
+            new byte[] { 23, 24, 25 }
         );
         final List<Float> expectedScores = dataVectors.stream()
-            .map(vector -> spaceType.getKnnVectorSimilarityFunction().compare(queryVector, vector))
+            .map(vector -> spaceType.getKnnVectorSimilarityFunction().compare(byteQueryVector, vector))
             .collect(Collectors.toList());
 
-        KNNFloatVectorValues values = mock(KNNFloatVectorValues.class);
+        KNNByteVectorValues values = mock(KNNByteVectorValues.class);
         when(values.getVector()).thenReturn(
             dataVectors.get(0),
             dataVectors.get(1),
@@ -85,6 +84,7 @@ public class VectorIdsKNNIteratorTests extends KNNTestCase {
             dataVectors.get(3),
             dataVectors.get(4)
         );
+
         // stub return value when nextDoc is called
         OngoingStubbing<Integer> stubbing = when(values.nextDoc());
         for (int i = 0; i < dataVectors.size(); i++) {
@@ -92,11 +92,12 @@ public class VectorIdsKNNIteratorTests extends KNNTestCase {
         }
         // set last return to be Integer.MAX_VALUE to represent no more docs
         stubbing.thenReturn(Integer.MAX_VALUE);
+
         // Execute and verify
-        VectorIdsKNNIterator iterator = new VectorIdsKNNIterator(queryVector, values, spaceType);
+        ByteVectorIdsExactKNNIterator iterator = new ByteVectorIdsExactKNNIterator(queryVector, values, spaceType);
         for (int i = 0; i < dataVectors.size(); i++) {
             assertEquals(i, iterator.nextDoc());
-            assertEquals(expectedScores.get(i), (Float) iterator.score());
+            assertEquals(expectedScores.get(i), iterator.score());
         }
         assertEquals(DocIdSetIterator.NO_MORE_DOCS, iterator.nextDoc());
         verify(values, never()).advance(anyInt());